time measurements as double not float

+++ examples/ios/OGLESGPGPUStillImage/OGLESGPGPUStillImage/RootViewController.mm
@@ -281,12 +281,12 @@ static NSArray *availableTestImages = [NSArray arrayWithObjects:
 #ifdef OGLES_GPGPU_BENCHMARK
     NSLog(@"Time measurements:");
     NSLog(@"---");
-    vector<float> timeMeasurements = gpgpuMngr->getTimeMeasurements();
-    for (vector<float>::iterator it = timeMeasurements.begin();
+    vector<double> timeMeasurements = gpgpuMngr->getTimeMeasurements();
+    for (vector<double>::iterator it = timeMeasurements.begin();
          it != timeMeasurements.end();
          ++it)
     {
-        NSLog(@"> %f ms", *it);
+        NSLog(@"> %0.9f ms", *it);
     }
     NSLog(@"---");
 #endif
